Exit coordinator as soon as the last tier exits with non normal reason
This way we can set the restart startegy to temporary and set
max_restarts back to 0, causing the whole flow to restart and
not only individual parts.

The issue with restarting part of the flow is that it would likely
be short-lived, as the consumers would likely cascade a whole failure,
or, even worse, only part of the flow would work.